Add an empty baremetal implementation
