+++ eventrouter/internal/eventrouter_baremetal.c
@@ -1,0 +1,1 @@
+#include "eventrouter.h"
